refactor: move Circle into SharedComponents
