+++ src/components/SearchResultsItem/SearchResultsSubjectIcon.tsx
@@ -1,8 +1,7 @@
 import { FC } from "react";
 
-import SubjectIcon from "../SubjectIcon";
-import Circle from "../Circle";
-
+import SubjectIcon from "@/components/SubjectIcon";
+import Circle from "@/components/SharedComponents/Circle";
 import Flex from "@/components/SharedComponents/Flex";
 
 export type SearchResultsSubjectIconProps = {
