+++ src/components/SharedComponents/Icon/GraphicCircleIcon.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 
 import Icon, { IconName } from ".";
 
-import Circle from "@/components/Circle";
+import Circle from "@/components/SharedComponents/Circle";
 import { OakColorName } from "@/styles/theme";
 import { FlexProps } from "@/components/SharedComponents/Flex";
